test: consolidate rename benchmark into rename_test.go
